로그인 기능 구현 코드 재배치
UserController에 있는 비즈니스 로직을 UserService 클래스로 옮겨서 비즈니스 로직을 Service클래스에 집중시켰다.
SessionUtil클래스는 Seesion의 key 값만 Static으로 가지고 있어서 SessionKeys로 이름을 바꿨고 세션에 로그인 아이디 등록하는 메소드는 로그인하는 목적에 맞게 login service로 옮겼다.
UserController에 있던 static LoginRequest, LoginResponse 객체는 코드 가독성을 높이고 일관된 코드를 가지게 하기 위해 따로 dto클래스로 만들었다.
login 컨트롤러의 반환값은 로그인 내용을 보내기에 충분한 LoginResponse의 객체로 바꿨다.
