+++ helparty/src/main/java/com/hamryt/helparty/dto/UserDto.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @ToString
 public class UserDto {
 
-    public enum Status{
+    public enum Status {
         DEFAULT, DELETE
     }
