+++ helparty/src/main/java/com/hamryt/helparty/service/UserService.java
@@ -1,10 +1,12 @@
 package com.hamryt.helparty.service;
 
+import com.hamryt.helparty.dto.LoginResponse;
 import com.hamryt.helparty.dto.UserDto;
+import javax.servlet.http.HttpSession;
 
 public interface UserService {
 
-    UserDto loginUser(String email, String password);
+    LoginResponse loginUser(String email, String password, HttpSession session);
 
     void insertUser(UserDto userDto);
