+++ helparty/src/main/java/com/hamryt/helparty/service/UserServiceImpl.java
@@ -1,9 +1,13 @@
 package com.hamryt.helparty.service;
 
+
+import com.hamryt.helparty.dto.LoginResponse;
 import com.hamryt.helparty.dto.UserDto;
 import com.hamryt.helparty.exception.EmailExistedException;
 import com.hamryt.helparty.exception.UnexpectedResultException;
 import com.hamryt.helparty.mapper.UserMapper;
+import com.hamryt.helparty.util.SessionKeys;
+import javax.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -19,12 +23,30 @@ public class UserServiceImpl implements UserService {
     private final Encryptor encryptor;
 
     @Override
-    public UserDto loginUser(String email, String password) {
+    public LoginResponse loginUser(String email, String password, HttpSession session) {
 
         String encryptPassword = encryptor.encrypt(password);
-        UserDto userDto = userMapper.findByIdAndPassword(email, encryptPassword);
+        UserDto userInfo = userMapper.findByIdAndPassword(email, encryptPassword);
 
-        return userDto;
+        LoginResponse loginResponse;
+
+        if (userInfo == null) {
+            loginResponse = LoginResponse.FAIL;
+
+        } else if (UserDto.Status.DEFAULT.equals(userInfo.getStatus())) {
+            loginResponse = LoginResponse.success(userInfo);
+            setLoginUserEmail(session, email);
+
+        } else {
+            log.error("Login Error");
+            throw new RuntimeException("Login Error");
+        }
+
+        return loginResponse;
+    }
+
+    public void setLoginUserEmail(HttpSession session, String email) {
+        session.setAttribute(SessionKeys.LOGIN_USER_EMAIL, email);
     }
 
     @Transactional
