+++ helparty/src/main/java/com/hamryt/helparty/util/SessionKeys.java
@@ -1,0 +1,11 @@
+package com.hamryt.helparty.util;
+
+public class SessionKeys {
+
+    private SessionKeys() {
+    }
+
+    public static final String LOGIN_USER_EMAIL = "LOGIN_USER_EMAIL";
+
+
+}
